Update Miniscript.PSBTExtensions
* Move every PSBT-related tests to NBitcoin.Tests.Miniscript
 * Add reference from NBitcoin.Tests.Miniscript to NBitcoin.Tests|TestFramework
 * Append TargetFrameworks `netcoreapp2.1;netcoreapp2.2` to TestFramework
* Rewrite PSBT.Finalize as an extension method in NBitcoin.Miniscript
