+++ NBitcoin.Miniscript.Tests/CSharp/NBitcoin.Miniscript.Tests.CSharp.csproj
@@ -15,6 +15,8 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\NBitcoin\NBitcoin.csproj" />
+    <ProjectReference Include="..\..\NBitcoin.TestFramework\NBitcoin.TestFramework.csproj" />
+    <ProjectReference Include="..\..\NBitcoin.Tests\NBitcoin.Tests.csproj" />
     <!-- <ProjectReference Include="..\..\NBitcoin.Miniscript\NBitcoin.Miniscript.fsproj" /> -->
   </ItemGroup>
   <ItemGroup>
@@ -24,4 +26,9 @@
     </Reference>
   </ItemGroup>
 
+  <ItemGroup>
+    <None Include="data\psbt.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
 </Project>
