+++ NBitcoin.TestFramework/NBitcoin.TestFramework.csproj
@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<VersionPrefix>1.6.35</VersionPrefix>
-		<TargetFrameworks>netstandard1.6;net452;netstandard2.0</TargetFrameworks>
+		<TargetFrameworks>netstandard1.6;net452;netstandard2.0;netcoreapp2.1;netcoreapp2.2</TargetFrameworks>
 		<TargetFrameworks Condition="'$(BuildingInsideVisualStudio)' == 'true'">netstandard2.0</TargetFrameworks>
     <TargetFrameworks Condition="'$(TargetFrameworkOverride)' != ''">$(TargetFrameworkOverride)</TargetFrameworks>
 		<AssemblyName>NBitcoin.TestFramework</AssemblyName>
