+++ NBitcoin/Properties/AssemblyInfo.cs
@@ -8,6 +8,8 @@ using System.Runtime.InteropServices;
 // associated with an assembly.
 [assembly: InternalsVisibleTo("NBitcoin.Tests")]
 [assembly: InternalsVisibleTo("NBitcoin.Altcoins")]
+[assembly: InternalsVisibleTo("NBitcoin.Miniscript")]
+[assembly: InternalsVisibleTo("NBitcoin.Miniscript.Tests.CSharp")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from
